# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/orders/{orderId}/cancel" : {
      "post" : {
        "tags" : [ "order-controller" ],
        "operationId" : "cancelOrder",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/internal/admin/orders/intervention/{orderId}/force-cancel" : {
      "post" : {
        "tags" : [ "admin-order-manual-controller" ],
        "operationId" : "forceCancelOrder",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "additionalProperties" : {
                  "type" : "string"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/admin/orders/intervention/{orderId}/cancel" : {
      "post" : {
        "tags" : [ "admin-order-manual-controller" ],
        "operationId" : "cancelOrder_1",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "additionalProperties" : {
                  "type" : "string"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/admin/orders/intervention/{orderId}/assign-driver" : {
      "post" : {
        "tags" : [ "admin-order-manual-controller" ],
        "operationId" : "assignDriver",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "additionalProperties" : {
                  "type" : "string"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/admin/orders/intervention/support-tickets/{ticketId}/resolve" : {
      "post" : {
        "tags" : [ "admin-order-manual-controller" ],
        "operationId" : "resolveSupportTicket",
        "parameters" : [ {
          "name" : "ticketId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "additionalProperties" : {
                  "type" : "string"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/internal/admin/orders/intervention" : {
      "get" : {
        "tags" : [ "admin-order-manual-controller" ],
        "operationId" : "getOrdersRequiringIntervention",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageOrder"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/admin/orders/intervention/support-tickets" : {
      "get" : {
        "tags" : [ "admin-order-manual-controller" ],
        "operationId" : "getOpenSupportTickets",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        }, {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "OPEN"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageSupportTicket"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ApiResponseVoid" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "type" : "object"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "SupportTicket" : {
        "required" : [ "createdAt", "customerId", "id", "orderId", "reason", "status" ],
        "type" : "object",
        "properties" : {
          "version" : {
            "type" : "integer",
            "format" : "int64"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "orderId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "customerId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "reason" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "OPEN", "IN_REVIEW", "RESOLVED", "REJECTED" ]
          },
          "resolutionNotes" : {
            "type" : "string"
          },
          "resolvedBy" : {
            "type" : "string",
            "format" : "uuid"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "resolvedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "chatSessionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "requestedRefundItems" : {
            "type" : "string"
          },
          "refundAmount" : {
            "type" : "number"
          },
          "restaurantComments" : {
            "type" : "string"
          },
          "riderComments" : {
            "type" : "string"
          }
        }
      },
...
      "ApiResponseString" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "PageableObject" : {
        "required" : [ "offset", "pageNumber", "pageSize", "paged", "unpaged" ],
        "type" : "object",
        "properties" : {
          "unpaged" : {
            "type" : "boolean"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "paged" : {
            "type" : "boolean"
          },
          "pageNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          },
          "offset" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "SortObject" : {
        "required" : [ "empty", "sorted", "unsorted" ],
        "type" : "object",
        "properties" : {
          "empty" : {
            "type" : "boolean"
          },
          "sorted" : {
            "type" : "boolean"
          },
          "unsorted" : {
            "type" : "boolean"
          }
        }
      },
...
      "Order" : {
        "required" : [ "createdAt", "customerId", "id", "restaurantId", "status", "totalAmount", "updatedAt" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "customerId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "customerName" : {
            "type" : "string"
          },
          "restaurantId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "restaurantName" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "CREATED", "PENDING_ACCEPTANCE", "AWAITING_DELAY_APPROVAL", "ACCEPTED", "PREPARING", "READY_FOR_PICKUP", "HANDED_OVER", "CANCELLED", "CANCELLED_BY_RESTAURANT" ]
          },
          "paymentMethod" : {
            "type" : "string",
            "enum" : [ "CARD", "UPI", "WALLET", "COD" ]
          },
          "cashCollectedAmount" : {
            "type" : "number"
          },
          "deliveryStatus" : {
            "type" : "string",
            "enum" : [ "PENDING", "SEARCHING_FOR_DRIVER", "MANUAL_INTERVENTION_REQUIRED", "ASSIGNED", "AT_RESTAURANT", "OUT_FOR_DELIVERY", "DELIVERED", "CANCELLED", "FAILED" ]
          },
          "paymentStatus" : {
            "type" : "string",
            "enum" : [ "INITIATED", "SUCCESS", "FAILED", "PENDING_COLLECTION", "COLLECTED", "PARTIALLY_REFUNDED", "REFUNDED", "REFUND_PENDING", "REFUND_FAILED" ]
          },
          "totalAmount" : {
            "type" : "number"
          },
          "itemTotal" : {
            "type" : "number"
          },
          "customerPlatformFee" : {
            "type" : "number"
          },
          "restaurantPlatformFee" : {
            "type" : "number"
          },
          "platformBonus" : {
            "type" : "number"
          },
          "restaurantDeliveryContribution" : {
            "type" : "number"
          },
          "restaurantPayout" : {
            "type" : "number"
          },
          "sgst" : {
            "type" : "number"
          },
          "cgst" : {
            "type" : "number"
          },
          "deliveryFee" : {
            "type" : "number"
          },
          "driverGrossPayout" : {
            "type" : "number"
          },
          "driverTaxes" : {
            "type" : "number"
          },
          "driverNetPayout" : {
            "type" : "number"
          },
          "quoteId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "rateBasePrice" : {
            "type" : "number"
          },
          "ratePerKm" : {
            "type" : "number"
          },
          "rateRestMaxContributionPercent" : {
            "type" : "number"
          },
          "rateFixedPlatformFee" : {
            "type" : "number"
          },
          "ratePlatformExcessCutPercent" : {
            "type" : "number"
          },
          "rateSgstPercent" : {
            "type" : "number"
          },
          "rateCgstPercent" : {
            "type" : "number"
          },
          "rateDeliverySgstPercent" : {
            "type" : "number"
          },
          "rateDeliveryCgstPercent" : {
            "type" : "number"
          },
          "distanceKm" : {
            "type" : "number"
          },
          "deliveryExecutiveId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "deliveryAddressId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "deliveryLat" : {
            "type" : "number",
            "format" : "double"
          },
          "deliveryLng" : {
            "type" : "number",
            "format" : "double"
          },
          "deliveryAddress" : {
            "type" : "string"
          },
          "pickupOtp" : {
            "type" : "string"
          },
          "otp" : {
            "type" : "string"
          },
          "estimatedPrepTimeMinutes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "estimatedCompletionTime" : {
            "type" : "integer",
            "format" : "int64"
          },
          "cancellationReason" : {
            "type" : "string"
          },
          "version" : {
            "type" : "integer",
            "format" : "int32"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "deliveredAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "PageOrder" : {
        "required" : [ "content", "empty", "first", "last", "number", "numberOfElements", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Order"
            }
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
      "PageSupportTicket" : {
        "required" : [ "content", "empty", "first", "last", "number", "numberOfElements", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SupportTicket"
            }
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
...
    }
  }
}